requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
xgboost>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.12.0
fredapi>=0.5.0
newsapi-python>=0.2.7
